requirements.txt
python-telegram-bot==20.7
openai
pytesseract
pandas
opencv-python
Pillow
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-auth==2.29.0
